+++ biopipen/scripts/scrna/RadarPlots.R
@@ -294,6 +294,7 @@ do_barplot_and_tests <- function(info, case, counts) {
         geom_errorbar(
             aes(ymin = mean_sd1, ymax = mean_sd2),
             width = 0.4,
+            linewidth = 0.8,
             position = position_dodge(0.9),
             color = "#333333"
         ) +
@@ -326,10 +327,21 @@ do_barplot_and_tests <- function(info, case, counts) {
                 group_by(!!sym(case$ident)) %>%
                 summarise(
                     comparison = paste0(pair, collapse = " - "),
+                    n = paste(as.list(table(!!sym(case$by)))[pair], collapse = "; "),
+                    mean = paste(
+                        (tibble(.frac, !!sym(case$by)) %>%
+                            group_by(!!sym(case$by)) %>%
+                            summarise(mean = mean(.frac)) %>%
+                            column_to_rownames(case$by) %>%
+                            t() %>%
+                            as.data.frame() %>%
+                            as.list())[pair] %>% unlist() %>% round(3),
+                        collapse = "; "
+                    ),
                     !!sym(paste0(case$test, "_pval")) := ifelse(
                         case$test == "wilcox",
-                        wilcox.test(.frac ~ !!sym(case$by))$p.value,
-                        t.test(.frac ~ !!sym(case$by))$p.value
+                        tryCatch(wilcox.test(.frac ~ !!sym(case$by))$p.value, error = function(e) NA),
+                        tryCatch(t.test(.frac ~ !!sym(case$by))$p.value, error = function(e) NA)
                     )
                 )
             test_results <- rbind(test_results, dat)
@@ -426,10 +438,11 @@ run_one_case <- function(casename) {
 
     # Get the counts
     if (!is.null(case$each)) {
-        counts <- meta %>% filter(!!sym(case$each) == case$each_value) %>% drop_na(!!sym(case$by))
+        counts <- meta %>% filter(!!sym(case$each) == case$each_value)
     } else {
-        counts <- meta %>% drop_na(!!sym(case$by))
+        counts <- meta
     }
+    counts <- counts %>% drop_na(!!sym(case$by)) %>% drop_na(!!sym(case$ident))
     do_radarplot(info, case, counts)
 
     if (!is.null(case$breakdown)) {